Move concludable compilation to LogicManager; Introduce ReasonerPlanner.planResolvableOdering
